Move abieosserializationobject back. Remove custom deserializer
